refactor: 旧 store/atoms.ts を新 presentation/store/ への再エクスポートに変換
既存フック（useHintMode, usePuzzleTimer, useCompletionOverlay, useVideoPlayback,
useGameState）のインポート先を新 presentation/store/ に直接切り替え。
store/atoms.ts は互換用の再エクスポートファイルに変換。

- ui-atoms.ts: completionOverlayVisibleAtom, videoPlaybackEnabledAtom, videoUrlAtom を個別定義に変更
- game-atoms.ts: puzzleStartTimeAtom を追加

全 6420 テストパス、ビルド成功
