+++ src/presentation/store/game-atoms.ts
@@ -16,6 +16,9 @@ export const selectedImageUrlAtom = atom<string | null>(null);
 /** 選択画像の元サイズ */
 export const selectedImageSizeAtom = atom<{ width: number; height: number } | null>(null);
 
+/** パズル開始時間（タイムスタンプ） */
+export const puzzleStartTimeAtom = atom<number | null>(null);
+
 /** 経過時間（秒） */
 export const gameElapsedTimeAtom = atom<number>(0);
