+++ src/store/atoms.ts
@@ -1,90 +1,55 @@
-import { atom } from 'jotai';
-import { GridPosition } from '../types/geometry';
+/**
+ * Jotai atoms — 後方互換の再エクスポート
+ *
+ * 全アトムは presentation/store/ に移動済み。
+ * 既存コードの互換性のため再エクスポートする。
+ */
 
 // 型の再エクスポート（後方互換性のため）
 export type { PuzzlePiece, ImageSize } from '../types/puzzle';
-import type { PuzzlePiece, ImageSize } from '../types/puzzle';
 
-// 初期値の定数
-const DEFAULT_PUZZLE_DIVISION = 4;
-const DEFAULT_PUZZLE_ELAPSED_TIME = 0;
+// ゲーム進行状態
+export {
+  selectedImageUrlAtom as imageUrlAtom,
+  selectedImageSizeAtom as originalImageSizeAtom,
+  puzzleStartTimeAtom,
+  gameElapsedTimeAtom as puzzleElapsedTimeAtom,
+  gameScoreAtom,
+  isBestScoreAtom,
+} from '../presentation/store/game-atoms';
 
-/**
- * アップロードされた画像のURLを管理するアトム
- */
-export const imageUrlAtom = atom<string | null>(null);
+// UI 状態
+export {
+  hintModeEnabledAtom,
+  completionOverlayVisibleAtom,
+  videoPlaybackEnabledAtom,
+  videoUrlAtom,
+  hintUsedAtom,
+} from '../presentation/store/ui-atoms';
 
-/**
- * 元の画像のサイズを管理するアトム
- */
-export const originalImageSizeAtom = atom<ImageSize | null>(null);
+// パズルボード状態 — 旧フック互換
+// 旧 usePuzzle.ts が個別 atom を使っているため、互換用に残す
+import { atom } from 'jotai';
+import { GridPosition } from '../types/geometry';
+import type { PuzzlePiece } from '../types/puzzle';
 
-/**
- * パズルの分割数（デフォルトは4x4=16ピース）を管理するアトム
- */
-export const puzzleDivisionAtom = atom<number>(DEFAULT_PUZZLE_DIVISION);
+/** パズルの分割数 */
+export const puzzleDivisionAtom = atom<number>(4);
 
-/**
- * パズルのピース配列を管理するアトム
- */
+/** パズルのピース配列 */
 export const puzzlePiecesAtom = atom<PuzzlePiece[]>([]);
 
-/**
- * 空白ピースの位置を管理するアトム
- */
+/** 空白ピースの位置 */
 export const emptyPiecePositionAtom = atom<GridPosition | null>(null);
 
-/**
- * パズルの開始時間（タイムスタンプ）を管理するアトム
- */
-export const puzzleStartTimeAtom = atom<number | null>(null);
-
-/**
- * パズルの経過時間（秒）を管理するアトム
- */
-export const puzzleElapsedTimeAtom = atom<number>(DEFAULT_PUZZLE_ELAPSED_TIME);
-
-/**
- * パズルが完成したかどうかを管理するアトム
- */
+/** パズルが完成したかどうか */
 export const puzzleCompletedAtom = atom<boolean>(false);
 
-/**
- * ヒントモードが有効かどうかを管理するアトム
- */
-export const hintModeEnabledAtom = atom<boolean>(false);
-
-/**
- * 完成オーバーレイの表示状態を管理するアトム
- */
-export const completionOverlayVisibleAtom = atom<boolean>(true);
-
-/**
- * 動画再生モードの状態を管理するアトム
- */
-export const videoPlaybackEnabledAtom = atom<boolean>(false);
-
-/**
- * 再生する動画のURLを管理するアトム
- */
-export const videoUrlAtom = atom<string | null>(null);
-
-/**
- * 現在の手数を管理するアトム
- */
+/** 現在の手数 */
 export const moveCountAtom = atom<number>(0);
 
-/**
- * シャッフル時の手数（= optimalMoves 基準値）を管理するアトム
- */
+/** シャッフル時の手数 */
 export const shuffleMovesAtom = atom<number>(0);
 
-/**
- * 正解位置にあるピースの割合（0〜100）を管理するアトム
- */
+/** 正解位置にあるピースの割合 */
 export const correctRateAtom = atom<number>(0);
-
-/**
- * ヒント使用フラグを管理するアトム
- */
-export const hintUsedAtom = atom<boolean>(false);
